Made changes to salary data type in Constructor

+++ src/main/java/com/olujobii/NonAcademicStaff.java
@@ -2,7 +2,7 @@ package com.olujobii;
 
 public class NonAcademicStaff extends Staff{
 
-    public NonAcademicStaff(String name, int age, String id, double salary) {
+    public NonAcademicStaff(String name, int age, String id, int salary) {
         super(name, age, id, salary);
     }
